+++ features/alerts/alertOnFishingBagDisabled.js
@@ -2,10 +2,11 @@ import settings from "../../settings";
 import { persistentData } from "../../data/data";
 import { GOLD, RED, WHITE } from "../../constants/formatting";
 import { getWorldName, hasFishingRodInHotbar, isInSkyblock } from "../../utils/playerState";
-import { OFF_SOUND_MODE } from "../../constants/sounds";
+import { MC_RANDOM_ORB_SOUND, OFF_SOUND_MODE } from "../../constants/sounds";
 import { getLore, isFishingHookActive, isInFishingWorld } from "../../utils/common";
 import { USE_BAITS_FROM_FISHING_BAG_DISABLED, USE_BAITS_FROM_FISHING_BAG_ENABLED } from "../../constants/triggers";
 import { registerIf } from "../../utils/registers";
+import { playMcSound } from "../../utils/sound";
 
 // Alert once after each world load, do not alert on each rod cast
 let isAlerted = false;
@@ -16,12 +17,12 @@ registerIf(
 );
 
 registerIf(
-    register('Chat', (event) => setFishingBagState(false)).setCriteria(USE_BAITS_FROM_FISHING_BAG_DISABLED),
+    register('Chat', (event) => setFishingBagState(false)).setCriteria(USE_BAITS_FROM_FISHING_BAG_DISABLED).setStart(),
     () => settings.alertOnFishingBagDisabled && isInSkyblock()
 );
 
 registerIf(
-    register('Chat', (event) => setFishingBagState(true)).setCriteria(USE_BAITS_FROM_FISHING_BAG_ENABLED),
+    register('Chat', (event) => setFishingBagState(true)).setCriteria(USE_BAITS_FROM_FISHING_BAG_ENABLED).setStart(),
     () => settings.alertOnFishingBagDisabled && isInSkyblock()
 );
 
@@ -59,7 +60,7 @@ function alertOnFishingBagDisabled() {
         Client.showTitle(`${RED}Enable fishing bag!`, '', 1, 25, 1);
     
         if (settings.soundMode !== OFF_SOUND_MODE) {
-            World.playSound('random.orb', 1, 1);
+            playMcSound(MC_RANDOM_ORB_SOUND);
         }
 
         const message = new TextComponent(`${GOLD}[FeeshNotifier] ${WHITE}Using baits from Fishing Bag is disabled. Click to open Fishing Bag!`).setClick("run_command", `/fb`);