+++ features/alerts/alertOnGoldenFish.js
@@ -1,13 +1,14 @@
 import settings from "../../settings";
 import * as triggers from '../../constants/triggers';
-import { OFF_SOUND_MODE } from "../../constants/sounds";
+import { MC_RANDOM_SPLASH_SOUND, OFF_SOUND_MODE } from "../../constants/sounds";
 import { GOLD, WHITE } from "../../constants/formatting";
 import { getWorldName, isInSkyblock } from "../../utils/playerState";
 import { registerIf } from "../../utils/registers";
 import { CRIMSON_ISLE } from "../../constants/areas";
+import { playMcSound } from "../../utils/sound";
 
 registerIf(
-	register("Chat", (event) => playAlertOnGoldenFish()).setCriteria(triggers.GOLDEN_FISH_MESSAGE),
+	register("Chat", (event) => playAlertOnGoldenFish()).setCriteria(triggers.GOLDEN_FISH_MESSAGE).setStart(),
 	() => settings.alertOnGoldenFishSpawned && isInSkyblock() && getWorldName() === CRIMSON_ISLE
 );
 
@@ -20,7 +21,7 @@ function playAlertOnGoldenFish() {
 		Client.showTitle(`${WHITE}Catch ${GOLD}Golden Fish`, '', 1, 30, 1);
 	
 		if (settings.soundMode !== OFF_SOUND_MODE) {
-            World.playSound('random.splash', 1, 1);
+            playMcSound(MC_RANDOM_SPLASH_SOUND);
         }
 	} catch (e) {
 		console.error(e);