+++ features/alerts/alertOnSpiritMaskUsed.js
@@ -1,12 +1,13 @@
 import settings from "../../settings";
 import * as triggers from '../../constants/triggers';
-import { OFF_SOUND_MODE } from "../../constants/sounds";
+import { MC_RANDOM_ORB_SOUND, OFF_SOUND_MODE } from "../../constants/sounds";
 import { DARK_PURPLE, GOLD, GREEN, WHITE, YELLOW } from "../../constants/formatting";
 import { isInSkyblock } from "../../utils/playerState";
 import { registerIf } from "../../utils/registers";
+import { playMcSound } from "../../utils/sound";
 
 registerIf(
-	register("Chat", (event) => playAlertOnSpiritMaskUsed()).setCriteria(triggers.SPIRIT_MASK_USED_MESSAGE),
+	register("Chat", (event) => playAlertOnSpiritMaskUsed()).setCriteria(triggers.SPIRIT_MASK_USED_MESSAGE).setStart(),
 	() => settings.alertOnSpiritMaskUsed && isInSkyblock()
 );
 
@@ -19,7 +20,7 @@ function playAlertOnSpiritMaskUsed() {
 		Client.showTitle(`${YELLOW}Spirit Mask used`, '', 1, 30, 1);
 	
 		if (settings.soundMode !== OFF_SOUND_MODE) {
-            World.playSound('random.orb', 1, 1);
+            playMcSound(MC_RANDOM_ORB_SOUND);
         }
 
 		if (settings.alertOnSpiritMaskBack) {
@@ -37,7 +38,7 @@ function playAlertOnSpiritMaskBack() {
 		Client.showTitle(`${GREEN}Spirit Mask ready`, '', 1, 30, 1);
 	
 		if (settings.soundMode !== OFF_SOUND_MODE) {
-            World.playSound('random.orb', 1, 1);
+            playMcSound(MC_RANDOM_ORB_SOUND);
         }
 	} catch (e) {
 		console.error(e);