+++ features/overlays/seaCreaturesCountAndTimeTracker.js
@@ -1,12 +1,12 @@
 import settings, { allOverlaysGui } from "../../settings";
-import { GOLD, RED, DARK_GRAY, WHITE, GRAY } from "../../constants/formatting";
+import { GOLD, RED, DARK_GRAY, WHITE, GRAY, BOLD } from "../../constants/formatting";
 import { TIMER_SOUND_SOURCE, OFF_SOUND_MODE } from "../../constants/sounds";
 import { ALL_SEA_CREATURES_NAMES } from "../../constants/seaCreatures";
 import { EntityArmorStand } from "../../constants/javaTypes";
 import { overlayCoordsData } from "../../data/overlayCoords";
 import { getWorldName, hasFishingRodInHotbar, isInHunterArmor, isInSkyblock } from "../../utils/playerState";
 import { CRIMSON_ISLE, CRYSTAL_HOLLOWS, HUB } from "../../constants/areas";
-import { createButtonsDisplay, toggleButtonsDisplay } from "../../utils/overlays";
+import { Overlay, OverlayTextLine, OverlayButtonLine, LEFT_CLICK_TYPE } from "../../utils/overlays";
 import { registerIf } from "../../utils/registers";
 import { isInFishingWorld } from "../../utils/common";
 
@@ -31,7 +31,7 @@ registerIf(
 );
 
 registerIf(
-    register('renderOverlay', () => renderCountOverlay()),
+    register('step', () => refreshOverlay()).setFps(2),
     () => settings.seaCreaturesCountOverlay && isInSkyblock() && isInFishingWorld(getWorldName())
 );
 
@@ -39,7 +39,9 @@ register("worldUnload", () => {
     resetSeaCreaturesCountAndTimer();
 });
 
-const buttonsDisplay = createButtonsDisplay(true, () => resetSeaCreaturesCountAndTimer(), false, null);
+const overlay = new Overlay(() => settings.seaCreaturesCountOverlay && isInSkyblock() && isInFishingWorld(getWorldName()))
+    .setPositionData(overlayCoordsData.seaCreaturesCountOverlay)
+    .setIsClickable(true);
 
 export function resetSeaCreaturesCountAndTimer() {
     startTime = null;
@@ -131,7 +133,9 @@ function alertOnSeaCreaturesTimerThreshold() {
     }
 }
 
-function renderCountOverlay() {
+function refreshOverlay() {
+    overlay.clear();
+
     if (!settings.seaCreaturesCountOverlay ||
         !mobsCount ||
         !startTime ||
@@ -141,7 +145,6 @@ function renderCountOverlay() {
         !hasFishingRodInHotbar() ||
         allOverlaysGui.isOpen()
     ) {
-        buttonsDisplay.hide();
         return;
     }
 
@@ -160,12 +163,12 @@ function renderCountOverlay() {
     const seaCreaturesColor = mobsCount >= getSeaCreaturesCountThreshold() ? RED : GOLD;
 
     const overlayText = `${seaCreaturesColor}${mobsCount} ${GRAY}${seaCreaturesText} ${DARK_GRAY}(${timerColor}${timerText}${DARK_GRAY})`;
-    const overlay = new Text(overlayText, overlayCoordsData.seaCreaturesCountOverlay.x, overlayCoordsData.seaCreaturesCountOverlay.y)
-        .setShadow(true)
-        .setScale(overlayCoordsData.seaCreaturesCountOverlay.scale);
-    overlay.draw();
-
-    toggleButtonsDisplay(buttonsDisplay, overlay, overlayCoordsData.seaCreaturesCountOverlay);
+    
+    overlay.addTextLine(new OverlayTextLine().setText(overlayText));
+    overlay.addButtonLine(new OverlayButtonLine()
+        .setText(`${RED}${BOLD}[Click to reset]`)
+        .setIsSmallerScale(true)
+        .setOnClick(LEFT_CLICK_TYPE, () => resetSeaCreaturesCountAndTimer()));
 }
 
 function getSeaCreaturesCountThreshold() {