+++ utils/sound.js
@@ -1,7 +1,11 @@
 import settings from "../settings";
-import { OFF_SOUND_MODE } from "../constants/sounds";
+import { MC_NOTE_PLING_SOUND, OFF_SOUND_MODE } from "../constants/sounds";
 
 let rareDropSoundLastPlayedAt = null;
+
+export function playMcSound(name, volume = 1, pitch = 1) {
+    World.playSound(name, volume, pitch);
+}
 
 // https://minecraft.fandom.com/wiki/Note_Block
 // Pitches extracted via soundPlay register, and mapped to the table above
@@ -19,8 +23,8 @@ export function playRareDropSound() {
 
     const volume = 0.8;
 
-    World.playSound('note.pling', volume, Math.pow(2, -9 / 12));
-    setTimeout(() => World.playSound('note.pling', volume, Math.pow(2, -4 / 12)), 200);
-    setTimeout(() => World.playSound('note.pling', volume, Math.pow(2, 1 / 12)), 400);
-    setTimeout(() => World.playSound('note.pling', volume, Math.pow(2, 3 / 12)), 600);
+    playMcSound(MC_NOTE_PLING_SOUND, volume, Math.pow(2, -9 / 12));
+    setTimeout(() => playMcSound(MC_NOTE_PLING_SOUND, volume, Math.pow(2, -4 / 12)), 200);
+    setTimeout(() => playMcSound(MC_NOTE_PLING_SOUND, volume, Math.pow(2, 1 / 12)), 400);
+    setTimeout(() => playMcSound(MC_NOTE_PLING_SOUND, volume, Math.pow(2, 3 / 12)), 600);
 }